typing & seperation
