+++ packages/cover-traffic/src/dashboard.ts
@@ -1,6 +1,8 @@
 import blessed from 'blessed'
 import contrib from 'blessed-contrib'
-import { main, State, findChannelsFrom, importance, totalChannelBalanceFor, findChannel, OpenChannels } from './ct'
+import { State, OpenChannels } from './state'
+import { findChannel, totalChannelBalanceFor, findChannelsFrom, importance } from './utils'
+import { main } from '.'
 import { privKeyToPeerId } from '@hoprnet/hopr-utils'
 import { PublicKey } from '@hoprnet/hopr-utils'
 import { BigNumber } from 'bignumber.js'
